Comment old benchmarks temporarily
